feat(WorkImages): add message for no images available in WorkImages component to enhance user experience refactor(WorkMetadata): simplify props by using hooks for data fetching instead of passing multiple props refactor(ArrangerWork): streamline data fetching by removing unnecessary hooks and props refactor(ExhibitionWork): optimize component by using work directly instead of workListItem.work
