+++ client/src/pages/ExhibitionWork.tsx
@@ -1,9 +1,5 @@
 import { type components } from "../types/api";
-import { useArranger } from "../hooks/arranger";
-import { useCategory } from "../hooks/category";
 import { useExhibition } from "../hooks/exhibition";
-import { useMaterials } from "../hooks/material";
-import { useSeason } from "../hooks/season";
 import { useExhibitionWorkListItem } from "../hooks/exhibition";
 import "./work.css";
 import { Link, useParams } from "react-router";
@@ -53,14 +49,10 @@ export default function ExhibitionWork() {
 
   const workListItem = useExhibitionWorkListItem(exhibitionId, workId);
   const work = workListItem?.work;
-  const arranger = useArranger(work?.arranger_id);
-  const category = useCategory(work?.category_id);
   const exhibition = useExhibition(work?.exhibition_id);
-  const materials = useMaterials(work?.material_ids);
-  const season = useSeason(work?.season_id);
 
   // TODO: ローディングと不正なアクセスを切り分けて表示する
-  if (!workListItem || !arranger || !category || !exhibition || !materials || !season) {
+  if (!workListItem || !work  || !exhibition) {
     return (
       <main>
         <h1>指定された作品は存在しません</h1>
@@ -72,15 +64,10 @@ export default function ExhibitionWork() {
     <>
       <main>
         <WorkHeading exhibition={exhibition} />
-        <WorkImages work={workListItem.work} />
+        <WorkImages work={work} />
         <AdjacentNavigation exhibitionId={exhibitionId} navigation={workListItem.navigation} />
         <WorkMetadata
-          arranger={arranger}
-          category={category}
-          exhibition={exhibition}
-          materials={materials}
-          season={season}
-          work={workListItem.work}
+          work={work}
         />
       </main>
     </>
